Separate types and nodes.
