+++ docs/classes/XmlValueNode.json
@@ -20,7 +20,7 @@
       ]
     },
     "description": "A node for a plain XML value (the value itself, not the tag that contains it). This node can have a value, and cannot have children.",
-    "sourceCode": "src/xml.ts"
+    "sourceCode": "src/lib/nodes.ts"
   },
   "sections": [
     {
